main

+++ src/main.cpp
@@ -8,7 +8,7 @@
 #define _DIV(left, right)   createNode(left, right, OPER, DIV)
 #define _MULT(left, right)  createNode(left, right, OPER, MULT)
 
-#define _NUM(value)         createNode(NULL, NULL, NUM, value)
+#define _NUM(value)         createNode(NULL, NULL, NUM, (double)value)
 
 #define _X                  createNode(NULL, NULL, VAR, 0)
 
@@ -16,7 +16,7 @@ int main(){
     BinExpTree* tree = NULL;
     binExpTreeCtor(&tree INIT_BET(tree));
 
-    tree->root = _DIV(_ADD(_X, _NUM(3.0)), _SUB(_NUM(1000.0), _NUM(7.0)));
+    tree->root = _DIV(_ADD(_X, _NUM(3)), _SUB(_NUM(1000), _NUM(7)));
     int nodes_amount = 0;
     countNodes(tree->root, &nodes_amount);
     tree->nodes_amount = nodes_amount;
